requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
openpyxl==3.1.2
fpdf2==2.7.9
